tests for ready
